+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix: Add link props to `SimilarProjectCard` component.
 - Fix: Pass `onRefreshClick` prop correctly to `SimilarProjectCard` component in
   `SimilarProjectsCard` component.
+- Fix: Set current index to 0 when `projects` props is updated in
+  `SimilarProjectsCard` component.
 
 ## [13.6.1] - 2017-08-03